Better plans project layout.
