+++ InstructionSets/x86/Decoder.cpp
@@ -6,7 +6,7 @@
 //  Copyright © 2021 Thomas Harte. All rights reserved.
 //
 
-#include "x86.hpp"
+#include "Decoder.hpp"
 
 #include <algorithm>
 #include <cassert>
@@ -17,7 +17,7 @@ using namespace CPU::Decoder::x86;
 // Only 8086 is suppoted for now.
 Decoder::Decoder(Model) {}
 
-std::pair<int, Instruction> Decoder::decode(const uint8_t *source, size_t length) {
+std::pair<int, CPU::Decoder::x86::Instruction> Decoder::decode(const uint8_t *source, size_t length) {
 	const uint8_t *const end = source + length;
 
 	// MARK: - Prefixes (if present) and the opcode.
